requirements.txt
boto3==1.9.101
dj-database-url==0.5.0
Django>=2.1.5
django-filter == 2.1.0
django-forms-bootstrap==3.1.0
django-materializecss-form==1.1.11
django-storages==1.6.6
djangorestframework==3.9.1
gunicorn==19.9.0
numpy==1.16.2
pandas==0.24.1
psycopg2==2.7.7
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
urllib3==1.24.1
